Initial commit
Began creating this tiny timecard.
Update timecard to allow date changing and download as csv. Got carried away and did everything on my little list and implemented back and next buttons next to the day picker. I think it's done - except that the download can potentially have the dates in the wrong order.